Revert "spearmint: Make botlib parse vectors in entity pairs as floats"
This reverts commit 8b5121c78abc21ec2a3e4b03c727b1acf7a40eba.

I changed it from double to float to match Game VM's entity spawn
logic and because my botlib ported to Game VM changed it to float
due to QVM sscanf not supporting %lf. I have since made QVM sscanf
support %lf and reverted this changed in mint-arena.

+++ code/botlib/be_aas_bspq3.c
@@ -248,13 +248,13 @@ int AAS_ValueForBSPEpairKey(int ent, char *key, char *value, int size)
 int AAS_VectorForBSPEpairKey(int ent, char *key, vec3_t v)
 {
 	char buf[MAX_EPAIRKEY];
-	float v1, v2, v3;
+	double v1, v2, v3;
 
 	VectorClear(v);
 	if (!AAS_ValueForBSPEpairKey(ent, key, buf, MAX_EPAIRKEY)) return qfalse;
-	//scanf into floats, then assign, so it is vec_t size independent
+	//scanf into doubles, then assign, so it is vec_t size independent
 	v1 = v2 = v3 = 0;
-	sscanf(buf, "%f %f %f", &v1, &v2, &v3);
+	sscanf(buf, "%lf %lf %lf", &v1, &v2, &v3);
 	v[0] = v1;
 	v[1] = v2;
 	v[2] = v3;
